fix(darkai): adopt size_t policy, RAII internals, remove unreachable code (-157 warnings via root-cause type policy, not suppression)
dbp_plugin_darkai (+ shared deps) /W4 families eliminated: C4701 5->0 (init at declaration), C4456/C4458/C4459 35->0 (rename/remove shadowing incl. Beacon::SetOther latent param shadow), C4189 63->0 (dead locals deleted after side-effect review), C4244/C4267 55->0 (size_t receivers; _findfirst intptr_t; sqrtf/fabsf for float math; DWORD kept at binary/Win32 boundaries with explicit casts), C4702 2->0. Build green, dbp_tests 886/886.

+++ Dark Basic Public Shared/DBPCompiler/Encryptor.cpp
@@ -31,7 +31,7 @@ bool CEncryptor::EncryptFileData(LPSTR filebuffer, DWORD filebuffersize, bool bE
 
 		// Key Index
 		DWORD dwKeyIndex=0;
-		DWORD dwKeyMax=strlen(pNewKey);
+		size_t dwKeyMax=strlen(pNewKey);
 
 		// Encrupt Data using Key (mess the file data up)
 		DWORD dwSpan = filebuffersize/1024;
